refactor: move func that makes input compatible with transformers to core module
That function is not for the utilities module, which includes user-facing functionalities.
